requirements.txt
openai
chromadb
langchain-openai
langchain_community
langgraph
pypdf
ipykernel
python-dotenv
fastapi
uvicorn
ragas